backend/pyproject.toml
[project]
name = "sisyphus-case-platform"
version = "0.1.0"
description = "AI-driven enterprise test case generation platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "asyncpg>=0.30",
    "psycopg2-binary>=2.9",
    "pydantic-settings>=2.7",
    "celery[redis]>=5.4",
    "redis>=5.2",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "httpx>=0.28",
    "python-multipart>=0.0.18",
    "minio>=7.2",
    "zhipuai>=2.1.5.20250825",
    "openai>=2.26.0",
    "python-docx>=1.1",
    "pypdf>=5.0",
    "markdown>=3.5",
    "qdrant-client>=1.12",
    "openpyxl>=3.1",
    "xmindparser>=1.2.2",
    "langsmith>=0.2",
    "langchain-core>=0.3",
    "langchain-openai>=0.2",
    "fastembed>=0.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "factory-boy>=3.3",
    "ruff>=0.8",
    "pyright>=1.1",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E", "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "SIM",
    "ASYNC",
]

[tool.ruff.lint.per-file-ignores]
"app/ai/prompts.py" = ["E501"]  # Long prompt strings are exempt from line-length limits

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
]
